Refactor the DNS resolver used to resolve internet names to avoid confusion with the new net address resolver mechanism
